refactor and cleaning
